pyproject.toml
[tool.poetry]
name = "fetchy"
version = "0.1.6"
description = "Minuscule docker images made trivial!"
authors = ["thomas <thomas.kluiters@gmail.com>"]
homepage = "https://github.com/ThomasKluiters/fetchy"
repository = "https://github.com/ThomasKluiters/fetchy"

[tool.poetry.dependencies]
python = "^3.6"
tqdm = "^4.32"
distro = "^1.4"
validators = "^0.13.0"
PyInquirer = "^1.0"
unix_ar = "^0.2.1"
docker = "^4.0"
cleo = "^0.7.5"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = { version = "*", allows-prereleases = true }
flake8 = "3.7.8"

[tool.poetry.scripts]
fetchy = 'fetchy.cli:main'
